% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sandwich.R
\name{.build_H_obs}
\alias{.build_H_obs}
\title{Build the observed information matrix}
\usage{
.build_H_obs(X, v, group, J, p, beta_prior_prec, theta_prior_prec)
}
\arguments{
\item{X}{Design matrix (N x p).}

\item{v}{Working weights (length N).}

\item{group}{Integer group indicator (1:J), length N.}

\item{J}{Number of groups.}

\item{p}{Number of fixed-effect parameters.}

\item{beta_prior_prec}{Scalar prior precision for beta.}

\item{theta_prior_prec}{Scalar prior precision for theta.}
}
\value{
Symmetric (p+J) x (p+J) matrix.
}
\description{
Constructs the (p+J) x (p+J) block matrix H_obs from working weights.
}
\keyword{internal}
