% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sandwich.R
\name{.build_V_sand}
\alias{.build_V_sand}
\title{Build the sandwich variance estimator}
\usage{
.build_V_sand(H_obs_inv, J_cluster)
}
\arguments{
\item{H_obs_inv}{Inverse of the observed information matrix (d x d).}

\item{J_cluster}{Clustered score outer product matrix (d x d).}
}
\value{
Symmetric (d x d) sandwich variance matrix.
}
\description{
Computes V_sand = H_obs_inv \%\emph{\% J_cluster \%}\% H_obs_inv and symmetrizes.
}
\keyword{internal}
