% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/der_correct.R
\name{as.matrix.svyder}
\alias{as.matrix.svyder}
\title{Extract Draws Matrix from a svyder Object}
\usage{
\method{as.matrix}{svyder}(x, ...)
}
\arguments{
\item{x}{A \code{svyder} object.}

\item{...}{Ignored.}
}
\value{
A numeric matrix of posterior draws (S x d).
}
\description{
Returns the corrected draws if available, otherwise the original draws.
This method allows \code{svyder} objects to be used wherever a numeric
matrix of posterior draws is expected.
}
\examples{
data(nsece_demo)
result <- der_diagnose(
  nsece_demo$draws,
  y = nsece_demo$y, X = nsece_demo$X,
  group = nsece_demo$group, weights = nsece_demo$weights,
  psu = nsece_demo$psu, family = "binomial",
  sigma_theta = nsece_demo$sigma_theta,
  param_types = nsece_demo$param_types
)
draws <- as.matrix(result)
dim(draws)

}
\seealso{
\code{\link[=der_correct]{der_correct()}} for applying the correction.

Other svyder-methods: 
\code{\link{glance.svyder}()},
\code{\link{is.svyder}()},
\code{\link{print.svyder}()},
\code{\link{summary.svyder}()},
\code{\link{tidy.svyder}()}
}
\concept{svyder-methods}
