% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sandwich.R
\name{.build_J_cluster}
\alias{.build_J_cluster}
\title{Build the clustered score outer product matrix}
\usage{
.build_J_cluster(X, r, psu, group, p, J)
}
\arguments{
\item{X}{Design matrix (N x p).}

\item{r}{Weighted residuals (length N).}

\item{psu}{Integer PSU indicator (1:G), length N.}

\item{group}{Integer group indicator (1:J), length N.}

\item{p}{Number of fixed-effect parameters.}

\item{J}{Number of groups.}
}
\value{
Symmetric (p+J) x (p+J) matrix.
}
\description{
Constructs J_cluster = sum_g s_g s_g^T where s_g is the score vector
summed within PSU g.
}
\keyword{internal}
