% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{is.svyder}
\alias{is.svyder}
\title{Test if an Object is a svyder Object}
\usage{
is.svyder(x)
}
\arguments{
\item{x}{An R object.}
}
\value{
Logical; \code{TRUE} if \code{x} inherits from class
\code{"svyder"}, \code{FALSE} otherwise.
}
\description{
Checks whether an R object inherits from class \code{"svyder"}.
}
\examples{
is.svyder(list())

data(nsece_demo)
result <- der_compute(
  nsece_demo$draws,
  y = nsece_demo$y, X = nsece_demo$X,
  group = nsece_demo$group, weights = nsece_demo$weights,
  psu = nsece_demo$psu, family = "binomial",
  sigma_theta = nsece_demo$sigma_theta,
  param_types = nsece_demo$param_types
)
is.svyder(result)

}
\seealso{
Other svyder-methods: 
\code{\link{as.matrix.svyder}()},
\code{\link{glance.svyder}()},
\code{\link{print.svyder}()},
\code{\link{summary.svyder}()},
\code{\link{tidy.svyder}()}
}
\concept{svyder-methods}
